+++ src/client/index.js
@@ -1,13 +1,14 @@
 import React from 'react';
-import createHistory from 'history/createBrowserHistory';
 import { hydrate } from 'react-dom';
 import { Provider } from 'react-redux';
-import { ConnectedRouter as Router, routerMiddleware } from 'react-router-redux';
+import { ConnectedRouter as Router, routerMiddleware } from 'connected-react-router';
+import { configureStore } from '../shared/store';
 import App from '../shared/App';
 import IntlProvider from '../shared/i18n/IntlProvider';
-import { configureStore } from '../shared/store';
+import createHistory from '../shared/store/history';
 
-const browserHistory = window.browserHistory || createHistory();
+const browserHistory = createHistory();
+
 const store =
     window.store ||
     configureStore({
@@ -31,8 +32,7 @@ if (process.env.NODE_ENV === 'development') {
         module.hot.accept();
     }
 
-    if (!window.store || !window.browserHistory) {
-        window.browserHistory = browserHistory;
+    if (!window.store) {
         window.store = store;
     }
 }