+++ src/shared/store/rootReducer.js
@@ -1,11 +1,13 @@
 import { combineReducers } from 'redux';
-import { routerReducer as router } from 'react-router-redux';
-
+import { connectRouter } from 'connected-react-router';
+import createHistory from './history';
 import app from './app/reducer';
+
+const history = createHistory();
 
 const rootReducer = combineReducers({
     app,
-    router,
+    router: connectRouter(history),
 });
 
 export default rootReducer;